+++ ouzel/graphics/ios/OpenGLView.mm
@@ -55,7 +55,8 @@
         CGPoint location = [touch locationInView:self];
 
         ouzel::sharedEngine->getInput()->touchBegin(reinterpret_cast<uint64_t>(touch),
-                                                    ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(location.x, location.y)));
+                                                    ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(static_cast<float>(location.x),
+                                                                                                                                       static_cast<float>(location.y))));
     }
 }
 
@@ -66,7 +67,8 @@
         CGPoint location = [touch locationInView:self];
 
         ouzel::sharedEngine->getInput()->touchMove(reinterpret_cast<uint64_t>(touch),
-                                                   ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(location.x, location.y)));
+                                                   ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(static_cast<float>(location.x),
+                                                                                                                                      static_cast<float>(location.y))));
     }
 }
 
@@ -77,7 +79,8 @@
         CGPoint location = [touch locationInView:self];
 
         ouzel::sharedEngine->getInput()->touchEnd(reinterpret_cast<uint64_t>(touch),
-                                                  ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(location.x, location.y)));
+                                                  ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(static_cast<float>(location.x),
+                                                                                                                                     static_cast<float>(location.y))));
     }
 }
 
@@ -88,7 +91,8 @@
         CGPoint location = [touch locationInView:self];
 
         ouzel::sharedEngine->getInput()->touchCancel(reinterpret_cast<uint64_t>(touch),
-                                                     ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(location.x, location.y)));
+                                                     ouzel::sharedEngine->getWindow()->convertWindowToNormalizedLocation(ouzel::Vector2(static_cast<float>(location.x),
+                                                                                                                                        static_cast<float>(location.y))));
     }
 }
